Wed, Aug 3, 2022, 10:36 PM +00:00

+++ psets/pset4/filter-less/helpers.c
@@ -17,9 +17,9 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width; j++)
         {
             // Get all 3 rgb colors
-            int red = image[i][j].rgbtRed;
-            int green = image[i][j].rgbtGreen;
-            int blue = image[i][j].rgbtBlue;
+            double red = image[i][j].rgbtRed;
+            double green = image[i][j].rgbtGreen;
+            double blue = image[i][j].rgbtBlue;
 
             // Calculate the average and round
             double average = (red + green + blue) / 3;
